Adds a missing "var" keyword to updateMarkers
The author forgot "var", so the "l" var is declared on the global scope. Apart from being wrong, it's dangerous for anyone using compiled/obfuscated JS, where it may be overwriting an obfuscated variable (common because it's only a single letter). I came across the bug when I noticed it overwriting what was supposed to be an alias for the window object in my compiled code.

Simply adding a var keyword redeclares it in the local scope, fixing the problem.